requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
psycopg2-binary==2.9.12
python-decouple==3.8
sqlparse==0.5.5
tzdata==2026.1
